Block payment from penalty details pages
