--- src/server/views/payment/pendingPayment.njk
@@ -1,44 +1,0 @@
-{% extends 'layouts/default.layout.njk' %}
-
-{% set pageTitle = t.site_title %}
-{% set pageBreadcrumbItems = [
-    { text: t.breadcrumbs.home, url: '/' },
-    { text: t.breadcrumbs.payment_code, url: '/payment-code' },
-    { text: t.breadcrumbs.penalty_details, url: '/payment-code/' + paymentCode },
-    { text: t.breadcrumbs.active_session }
-  ] 
-%}
-
-{% set paid = true if status == 'PAID' else false %}
-{% set statusClass = 'confirmed' if paid else 'unconfirmed' %}
-
-{% block content %}
-  
-  {% call components.gridRow() %}
-    {% call components.columnTwoThirds() %}
-      {{ components.heading(text=t.payment_pending_page.title, tag='h1', size='xlarge') }}
-      <div class="pending-payment-info-container">
-        {{ components.notice(text=t.payment_pending_page.pending_notice) }}
-      </div>
-      <div class="pending-payment-button-container">
-        {{ components.button(text=t.payment_pending_page.pay_button, type="alert", disabled="true") }}
-        {{ components.button(text=t.payment_pending_page.go_back, url=cancelUrl, start='true', type="info") }}
-      </div>
-    {%- endcall %}
-    {% call components.columnOneThird() %}
-      <aside class="govuk-related-items" role="complementary">
-        {{ components.heading(text=t.home.language_guidance_header + ':', tag='h3', size='small', id='language-links-header' )}}
-        <nav role="navigation" aria-labelledby="language-links-header">
-          <ul class="font-xsmall">
-            <li>{{ components.link(text='Cymraeg', url='?clang=cy') if clang != 'cy' else components.link(text='English', url='?clang=en') }}</li>
-            <li>{{ components.link(text='Deutsch', url='?clang=de') if clang != 'de' else components.link(text='English', url='?clang=en') }}</li>
-            <li>{{ components.link(text='Español', url='?clang=es') if clang != 'es' else components.link(text='English', url='?clang=en') }}</li>
-            <li>{{ components.link(text='Français', url='?clang=fr') if clang != 'fr' else components.link(text='English', url='?clang=en') }}</li>
-            <li>{{ components.link(text='Polski', url='?clang=pl') if clang != 'pl' else components.link(text='English', url='?clang=en') }}</li>
-          </ul>
-        </nav>
-      </aside>
-    {%- endcall %}
-  {%- endcall %}
-  
-{% endblock %}
